+++ scrapper/src/main/java/edu/java/client/dto/QuestionInformation.java
@@ -1,4 +1,4 @@
-package edu.java.dto;
+package edu.java.client.dto;
 
 import com.fasterxml.jackson.annotation.JsonSetter;
 import edu.java.utils.TimeManager;
